requirements.txt
customtkinter==5.2.1
ultralytics==8.0.227
opencv-python==4.8.1.78
numpy==1.26.2
pillow==10.1.0
tk==0.1.0